+++ pdscalibdata/trunk/src/CalibParsCenterV1.cpp
@@ -44,7 +44,7 @@ CalibParsCenterV1::CalibParsCenterV1 ()
 
 //----------------
 
-CalibParsCenterV1::CalibParsCenterV1( const std::vector<float> v_parameters )
+CalibParsCenterV1::CalibParsCenterV1( const std::vector<double> v_parameters )
 {
     if (v_parameters.size() != NUMBER_OF_PARAMETERS) {
         WithMsgLog("CalibParsCenterV1", error, str) {
@@ -55,7 +55,7 @@ CalibParsCenterV1::CalibParsCenterV1( const std::vector<float> v_parameters )
         abort();
     }
     size_t NPars = NQuad * NSect;
-    size_t arr_size = sizeof( float ) * v_parameters.size()/3;
+    size_t arr_size = sizeof( double ) * v_parameters.size()/3;
     memcpy( &m_center_x, &v_parameters[0],       arr_size );
     memcpy( &m_center_y, &v_parameters[NPars],   arr_size );
     memcpy( &m_center_z, &v_parameters[NPars*2], arr_size );
