+++ pdscalibdata/trunk/src/CalibParsOffsetV1.cpp
@@ -44,7 +44,7 @@ CalibParsOffsetV1::CalibParsOffsetV1 ()
 
 //----------------
 
-CalibParsOffsetV1::CalibParsOffsetV1( const std::vector<float> v_parameters )
+CalibParsOffsetV1::CalibParsOffsetV1( const std::vector<double> v_parameters )
 {
     if (v_parameters.size() != NUMBER_OF_PARAMETERS) {
         WithMsgLog("CalibParsOffsetV1", error, str) {
@@ -55,7 +55,7 @@ CalibParsOffsetV1::CalibParsOffsetV1( const std::vector<float> v_parameters )
         abort();
     }
     size_t NPars    = NQuad;
-    size_t arr_size = sizeof( float ) * v_parameters.size()/3;
+    size_t arr_size = sizeof( double ) * v_parameters.size()/3;
     memcpy( &m_offset_x, &v_parameters[0],       arr_size );
     memcpy( &m_offset_y, &v_parameters[NPars],   arr_size );
     memcpy( &m_offset_z, &v_parameters[NPars*2], arr_size );
